Testing an eCommerce Site with Selenium
In this Selenium project, we developed the Page Object Model framework and used it for buying a product on Flipkart. We tested the site for a search query like iPhone with the specific phone memory size requirement. The product was to be added to the cart, ensuring that shipping would be done to the entered address, and the site should proceed to the checkout page.